requirements.txt
Flask>=1.1.2
Frozen-Flask>=0.18
numpy>=1.19.2
openreview-py==1.1.0
pandas>=1.0.5
tqdm>=4.54.0